// frontend/src/Types/CompanyIncomeStatement.d.ts
export interface CompanyIncomeStatement {
    "date": string;
    "symbol": string;
    "reportedCurrency": string;
    "cik": string;
    "filingDate": string;
    "acceptedDate": string;
    "fiscalYear": string;
    "period": string;
    "revenue": number;
    "costOfRevenue": number;
    "grossProfit": number;
    "researchAndDevelopmentExpenses": number;
    "generalAndAdministrativeExpenses": number;
    "sellingAndMarketingExpenses": number;
    "sellingGeneralAndAdministrativeExpenses": number;
    "otherExpenses": number;
    "operatingExpenses": number;
    "costAndExpenses": number;
    "netInterestIncome": number;
    "interestIncome": number;
    "interestExpense": number;
    "depreciationAndAmortization": number;
    "ebitda": number;
    "ebit": number;
    "nonOperatingIncomeExcludingInterest": number;
    "operatingIncome": number;
    "totalOtherIncomeExpensesNet": number;
    "incomeBeforeTax": number;
    "incomeTaxExpense": number;
    "netIncomeFromContinuingOperations": number;
    "netIncomeFromDiscontinuedOperations": number;
    "otherAdjustmentsToNetIncome": number;
    "netIncome": number;
    "netIncomeDeductions": number;
    "bottomLineNetIncome": number;
    "eps": number;
    "epsDiluted": number;
    "weightedAverageShsOut": number;
    "weightedAverageShsOutDil": number;
}